pyproject.toml
[tool.poetry]
name = "repl_python3_scum-and-villainy"
version = "0.1.0"
description = ""
authors = ["EricSemrow <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
replit = "^3.2.4"
discord = "^1.7.3"
Flask = "^2.0.1"
d20 = "^1.1.2"
firebase-admin = "^5.0.2"
google-api-python-client = "^2.19.1"
google-auth-oauthlib = "^0.4.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
